Added App icon and unit tests. Fixed bugs.

+++ app/src/main/java/com/olq/piggifyme/data/database/Triplet.kt
@@ -6,9 +6,9 @@ package com.olq.piggifyme.data.database
 class Triplet(map: MutableMap<String, Any?>) {
     var type: String by map
     var source: String by map
-    var value: Int by map
+    var value: Long by map
 
-    constructor(type: String, source: String, value: Int) : this(HashMap()) {
+    constructor(type: String, source: String, value: Long) : this(HashMap()) {
         this.type = type
         this.source = source
         this.value = value
